// src/app/models/park.ts
export interface Park {
    id: number;
    park_name: string;
    park_abbreviation: string;
    park_img_url?: string;
    park_img_uri?: string;
    park_address: string;
    park_city: string;
    park_state: string;
    park_zip_code: number;
    park_latitude: number;
    park_longitude: number;
}